fix: different abi processing
If a different ABI was specified and another one was already processed,
it was skipped. Now it correctly detects new ABIs.

As a side effect, when no ABI is specified and the package only
supports some of them, the app will always be processed as currently
there is no way to tell which ABIs are supported per package